modif error compilation and makefile

+++ ft_memcpy.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   memcpy.c                                           :+:      :+:    :+:   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Yohan.H <yhurtebize58@gmail.com>           +#+  +:+       +#+        */
+/*   By: yohan.h <yohan.h@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/20 23:35:45 by Yohan.H           #+#    #+#             */
-/*   Updated: 2024/03/20 23:35:48 by Yohan.H          ###   ########.fr       */
+/*   Updated: 2024/04/15 12:15:40 by yohan.h          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t size)
 {
 	unsigned char	*cpy_dest;
 	unsigned char	*cpy_src;
-	int				i;
+	size_t          i;
 
 	cpy_dest = (unsigned char *)dest;
 	cpy_src = (unsigned char *)src;
